+++ tests/bootstrap.php
@@ -1,2 +1,3 @@
 <?php
 // for backward compatible
+error_reporting(E_ALL);