+++ tests/unit/Badoo/LiveProfilerUI/Pages/FlameGraphPageTest.php
@@ -205,14 +205,14 @@ class FlameGraphPageTest extends \unit\Badoo\BaseTestCase
             ->disableOriginalConstructor()
             ->setMethods(['getValue'])
             ->getMock();
-        $MethodTreeMock->method('getValue')->willReturn(200);
+        $MethodTreeMock->method('getValue')->willReturn(300);
         $data = [];
         for ($i = 0; $i < 3001; $i++) {
             $data[$i] = $MethodTreeMock;
         }
         return [
-            ['data' => [], 'expected' => 100],
-            ['data' => $data, 'expected' => 200],
+            ['data' => [], 'expected' => 300],
+            ['data' => $data, 'expected' => 300],
         ];
     }