+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 There are next changes:
 
+## 1.1.2
+
+There are next changes:
+
+- method list page was improved
+- data on method list and tree pages are sorted now by the first param
+- keep a graph period when go to a parent or child method
+
+## 1.1.1
+
+There are next changes:
+
+- added a flame graph by params diff
+- added a link to flame graph in the diff interface
+
 ## 1.1.0
 
 There are next changes: